+++ np.Shared/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 See [Shared Plugin's README](https://github.com/NotePlan/plugins/blob/main/np.Shared/README.md) for details on this plugin.
 
+## [0.5.7] - @dwertheimer
 ## [0.5.6] - @dwertheimer
 ## [0.5.5] - @dwertheimer
 ## [0.5.4] - @dwertheimer